% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBISW.R
\docType{data}
\name{FBISW}
\alias{FBISW}
\title{FBI Smith \& Wesson data}
\format{A data frame with 9453 rows and 6 columns of pairwise comparison data generated using the \code{cartridge3D} package.}
\usage{
FBISW
}
\description{
FBI Smith \& Wesson data
}
\keyword{datasets}
