% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyData.R
\name{genPairwise}
\alias{genPairwise}
\title{Generate pairwise comparisons}
\usage{
genPairwise(datasetName, labelColNum)
}
\arguments{
\item{datasetName}{name of data that is loaded in R e.g. AlonDS}

\item{labelColNum}{column number with cluster label}
}
\value{
data frame with all choose(n, 2) comparisons. Three columns: `hash1`
  is the row number in the original data set of the first item in the
  comparison, `hash2` is the second, and `match` is the ground truth for
  whether the pair is in the same cluster.
}
\description{
Generate pairwise comparisons
}
