% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyData.R
\name{genSimDiff}
\alias{genSimDiff}
\title{Generate either correlation, l1 or l2 distance, from feature(s)}
\usage{
genSimDiff(datasetName, featureCols, allPairwise, pairColNums, measure)
}
\arguments{
\item{datasetName}{name of data containing individual information}

\item{featureCols}{column(s) containing features that correlation, l1 or l2
distance are to be computed for}

\item{allPairwise}{name of data frame containing all pairwise comparisons.
This needs to have at least two columns, one representing the first item in
the comparison, and one representing the second item.}

\item{pairColNums}{vector of length 2 indicating the column numbers of 1.
item 1 in comparison, 2. item 2 in comparison}

\item{measure}{either "correlation", "l2dist" or "l1dist"}
}
\value{
allPairwise with an additional column, either "correlation", "l2dist"
  or "l1dist"
}
\description{
Generate either correlation, l1 or l2 distance, from feature(s)
}
