% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyData.R
\name{getMetrics}
\alias{getMetrics}
\title{Generate four evaluation metrics from pairwise comparisons}
\usage{
getMetrics(allPairwise, pairColNums, matchColNum, distSimCol, linkage,
  dist = TRUE)
}
\arguments{
\item{allPairwise}{name of data frame containing all pairwise comparisons.
This needs to have at least four columns, one representing the first item
in the comparison, one representing the second item, one representing the
true match/non-match status, and the last representing a distance or
similarity metric. These are enumerated in the next three parameters.}

\item{pairColNums}{vector of length 2 indicating the column numbers in
`allPairwise` of 1. item 1 in comparison, 2. item 2 in comparison}

\item{matchColNum}{column number of column in `allPairwise` indicating true
match/non-match status}

\item{distSimCol}{name of column in `allPairwise` indicating distances or
similarities, input as character, e.g. "l2dist". If this is a similarity
and not a difference, input `dist` parameter to be FALSE. If a similarity
measure is used, distance will be calcualted as 1 - similarity.}

\item{linkage}{one of "single", "complete", "average", "centroid", "minimax"}

\item{dist}{is `distSimCol` a distance or similarity measure? Default TRUE,
i.e. distance measure}
}
\value{
outMetrics, a data frame with each row representing a clustering. For
  a data set with $n$ items, there will be $n$ rows. Columns are the four
  evaluation metrics, `maxMinimax`, `misClass`, `precision` and `recall`.
}
\description{
This function runs hierarchical linkage using one of five linkage methods:
single linkage, complete linkage, average linkage, centroid linkage and
minimax linkage. For a data set with $n$ items, it is possible to get
clusterings of sizes 1 through n. For each cluster size, we compute four
evaluation metrics: 1. maximum minimax radius (see Bien et al. 2011), 2.
misclassification rate, 3. precision, 4. recall.
}
