% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forPlotting.R
\name{loadDataset}
\alias{loadDataset}
\title{Load evaluation metrics for all possible clusters k of a given data set}
\usage{
loadDataset(datasetName)
}
\arguments{
\item{datasetName}{name of the orginal data set}
}
\value{
data set with the following fields: `numClusters` `maxMinimax`  `misClass` `precision` `recall`  `Linkage`
}
\description{
Note that the outMetrics/ folder should already be created with the associated RData created from the `getMetrics()` function
}
