% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forPlotting.R
\name{plotResultsGG_base}
\alias{plotResultsGG_base}
\title{Create evaluation ggplot graphics for varying linkage results}
\usage{
plotResultsGG_base(
  single = NULL,
  complete = NULL,
  average = NULL,
  centroid = NULL,
  minimax = NULL,
  correlation = FALSE,
  write_plot = FALSE,
  plot_type = c("all", "minimax", "misclass", "pr")
)
}
\arguments{
\item{single}{name of of preloaded outmetrics for single linkage}

\item{complete}{name of of preloaded outmetrics for complete linkage}

\item{average}{name of of preloaded outmetrics for average linkage}

\item{centroid}{name of of preloaded outmetrics for centroid linkage}

\item{minimax}{name of of preloaded outmetrics for minimax linkage}

\item{correlation}{if TRUE, set y limits in max minimax graphic to be [0, 1]}

\item{write_plot}{if TRUE, write the graph to the working directory using ggsave()}

\item{plot_type}{which evaluation graph you want to produce}
}
\value{
evaluation graph, as specified in plot_type
}
\description{
Create evaluation ggplot graphics for varying linkage results
}
