% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forEvaluation.R
\name{precisionRecall}
\alias{precisionRecall}
\title{Calculate precision and recall given pairs, model prediction and true match
status}
\usage{
precisionRecall(allPairwise, linkCol, matchColNum)
}
\arguments{
\item{allPairwise}{name of data frame containing all pairwise comparisons.
This needs to have at least two columns, one representing whether the pair
is linked in the given clustering, and one representing the true
match/non-match status. These are enumerated in the next two parameters.}

\item{linkCol}{name of column in `allPairwise` with links, input as
character, e.g. "minimax0.4"}

\item{matchColNum}{column number of column in `allPairwise` indicating true
match/non-match status}
}
\value{
list with two items, `precision` and `recall`
}
\description{
Calculate precision and recall given pairs, model prediction and true match
status
}
