% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedDups.R
\docType{data}
\name{removedDups}
\alias{removedDups}
\title{NBIDE data}
\format{A data frame with 10296 rows and 6 columns of pairwise comparison data generated using the \code{cartridge3D} package.}
\usage{
removedDups
}
\description{
NBIDE data
}
\keyword{datasets}
