% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singh2002.R
\docType{data}
\name{singh2002}
\alias{singh2002}
\title{Prostate data set}
\format{A 102 x 6034 matrix, with columns 1 through 6033 gene expressions and
  the last column the class label.}
\source{
\url{ftp://stat.ethz.ch/Manuscripts/dettling/prostate.rda}
}
\usage{
singh2002
}
\description{
Prostate data from Singh 2002, see
\url{https://stat.ethz.ch/~dettling/bagboost.html}. Code used to produce this
is \code{singh2002 <- cbind(prostate.x, prostate.y); dimnames(singh2002) <-
NULL}.
}
\keyword{datasets}
