% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forPlotting.R
\name{plotResultsGG}
\alias{plotResultsGG}
\title{Create evaluation ggplot graphics for linkage results}
\usage{
plotResultsGG(
  datasetName,
  correlation = FALSE,
  write_plot = FALSE,
  plot_type = c("all", "minimax", "misclass", "pr")
)
}
\arguments{
\item{datasetName}{name of the orginal data set; to use this parameter you must have a folder called "outMetrics/", with files for each linkage method of the format: "datasetName_single.Rdata"}

\item{correlation}{if TRUE, set y limits in max minimax graphic to be [0, 1]}

\item{write_plot}{if TRUE, write the graph to the working directory using ggsave()}

\item{plot_type}{which evaluation graph you want to produce}
}
\value{
evaluation graph, as specified in plot_type
}
\description{
Create evaluation ggplot graphics for linkage results
}
